refactor: move instrumentation RenderProfile enum to its own module
This keeps this as a nice light-weight import that can be used by the
click CLI, also helping to keep potentially user-facing data boundaries
isolated.
